Move log file download to ValidationController
